buncha stuff

+++ preallocate_source.cpp
@@ -5,6 +5,9 @@
 #include <iostream>
 #include <vector>
 
+#include <stdlib.h>
+
+
 void usage(){
     std::cout << "preallocate N\n\nWhere N is amount of memory in MB to fill with objects, independently allocated on heap" << std::endl;
 }
@@ -21,7 +24,7 @@ int main(int argc, char* argv[]){
     long MB_IN_BYTES(1024*1024);
 
     // do some math
-    int n_objects = (mb_to_fill*MB_IN_BYTES) / sizeof(HeapObject);
+    long long n_objects = (mb_to_fill*MB_IN_BYTES) / sizeof(HeapObject);
 
 
     std::cout << "Filling " << mb_to_fill << " MB with " << n_objects << " objects of " << sizeof(HeapObject) << " bytes in size" << std::endl;
@@ -31,7 +34,7 @@ int main(int argc, char* argv[]){
     
     // do an operation on them
     bool resp(false);
-    for (int i = 0; i < n_objects; i++){
+    for (long long i = 0; i < n_objects; i++){
         for (int let_idx = 0; let_idx < 10; let_idx++){
             resp = objects->frobnicate(LETTERS[let_idx]);
         }
